requirements.txt
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.20.13
botocore==1.23.13
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
ecdsa==0.17.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
jmespath==0.10.0
numpy==1.21.4
pandas==1.3.4
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2021.3
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0

pytest~=7.1.2
python-dotenv~=1.0.0
alembic~=1.12.0
redis~=4.6.0